+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-#  (2020-08-27)
+#  (2020-08-28)
 
 
 ### Features
 
 - About description component ([4a45235](https://github.com/Exlabs/value-driven-templates/commit/4a45235f3c3411154403c5f1291df0d4a1fba25a))
-- Contribution Description ([3b1ee69](https://github.com/Exlabs/value-driven-templates/commit/3b1ee694a6aff944d5fbd55b838ea32f1f9d1fa4))
+- Contribution Description ([35f4ba5](https://github.com/Exlabs/value-driven-templates/commit/35f4ba530238cb64b7b5d1272637f55bae76e13a))
 - Contribution step ([61a3280](https://github.com/Exlabs/value-driven-templates/commit/61a3280427d096a01ebb83d31743d04b6b8f7f9e))
 - Footer component ([b10f707](https://github.com/Exlabs/value-driven-templates/commit/b10f707e635e773ee90fbb0a938abac715fc23e5))
 - Header about component ([fba1867](https://github.com/Exlabs/value-driven-templates/commit/fba18675d68e81e4e6a3e56131c13976435407f6))
@@ -25,6 +25,11 @@
 - Deployment improvement. Added static files prefix. ([653750b](https://github.com/Exlabs/value-driven-templates/commit/653750b2629faa9adbffb5e0557a18f3ad409258))
 - Empty thumbnail list improvement ([2148cb3](https://github.com/Exlabs/value-driven-templates/commit/2148cb37f7b1c79065779ffa990a7402b0d93227))
 - Fixed sticky nav jumping on mobile ([f0d3f87](https://github.com/Exlabs/value-driven-templates/commit/f0d3f87e6cd8704b03de76a276cd0fd237c95f78))
+
+
+### UI Updates
+
+- Styles improvements ([da297bb](https://github.com/Exlabs/value-driven-templates/commit/da297bb10cbdb2a7f50217c959022d47e08de237))
 
 
 ### Documentation